ref(grouping): Split up fingerprinting module (#98216)
This splits up the fingerprinting module both because a) it was too long to easily work with, and b) then its structure will match that of the enhancements module. No code changes, just moving stuff around.
